+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Improve performance of task settings setup
 - Improve raised errors messages
 - Removed the `Freezer` module and moved the logic into the executor `freeze_execution!`
+- Remove passing the task to callback signature
 
 ## [1.8.0] - 2025-09-22