+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Add option to output failure backtraces
 - Add option to handle exceptions for non bang exceptions
 - Add durability (retries) to execution
+- Added `to_h` hash coercion
 
 ### Changes
 - Clean halt backtrace frames