+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [UNRELEASED]
 
-## [1.9.0] - 2025-10-01
+## [1.9.0] - 2025-10-20
 
 ### Added
 - Add `transform` option to attributes